linked basket to product
Combined the prototype basket with products page

added dynamic logout button

+++ CMS/addProduct.php
@@ -14,8 +14,8 @@ $collection = $db->products;
 
 $productID = newID($collection);
 $productName= filter_input(INPUT_POST, 'name', FILTER_SANITIZE_STRING);
-$productPrice = filter_input(INPUT_POST, 'price', FILTER_SANITIZE_STRING);
-$productQuantity = filter_input(INPUT_POST, 'quantity', FILTER_SANITIZE_STRING);
+$productPrice = (int) filter_input(INPUT_POST, 'price', FILTER_SANITIZE_STRING);
+$productQuantity = (int) filter_input(INPUT_POST, 'quantity', FILTER_SANITIZE_STRING);
 $url = filter_input(INPUT_POST, 'url', FILTER_SANITIZE_STRING);
 $keyword = filter_input(INPUT_POST, 'keyword', FILTER_SANITIZE_STRING);
